No avatar image should be just that
I honestly don't understand the number that's shown instead of an avatar image if it's disabled in the settings. The code suggests it's a post count but in my testing it seemed wrong?

Regardless, the setting of disabling avatars should do just that and not add something else.

Perhaps add another possible setting to replace the avatar with some post count or something else... and then just an else if in the code above.