Fixed auto load of hashover/pdo.php file
I've implemented hashover with mysql database. When the library reads a comment it fires this error: "HashOver PDO.php file could not be included!" it happen because it try to load the file PDO.php that not exists. I've fixed the error by putting the slash in the "PDO::" calls