Move secrets configuration out of code and into ini file with template
Previously, users needed to edit the secrets.php file
before they could use HashOver. This requirement posed a
number of problems, such as a needlessly-complicated
upgrade procedure.

This commit moves the configuration to an .ini file in the
config/ directory. A secrets.ini is not included - instead,
there is a secrets.ini.sample file, which users can copy to
secrets.ini before editing. This allows simpler upgrades by
means of a 'git pull' or 'tar xf ...'.

The .ini file format was chosen as it is directly supported
by the PHP core (as INI support is needed by PHP itself to
read php.ini), allows comments (unlike JSON), and has a
simple and obvious syntax. Moving the file to be edited away
from backend/classes/ and into config/ also makes it more
discoverable.