Move root directory setting to secrets.ini
Though not really a "secret", it cannot be set in the .json
file (as the administration interface can't work without a
correct HTTP root), and a PHP define is ineffective because
it needs to be defined for every PHP entry point (i.e. not
just comments.php, but also comments-ajax.php etc.).