Removed element cache
Once upon a time this provided a performance improvement, nowadays it is negligible. The element cache has caused problems and confusion multiple times, especially when the code relies on cached elements that may no longer exist on the page. Adding checks to ensure a cached element is still on the page before using it degrades performance worse than the original performance improvement gained from implementing the cache in the first place.

For these reasons and the fact that the cache increases memory use, it is being removed entirely.