Removed `HashOver.{}.parent` property
In preparation for supporting multiple threads/instances on a single page, this property is being removed.

This a single HashOver instance, this is fine, but with multiple instances the approach of storing the instance as a `parent` property no longer works, as it is pass by reference and always points to be last instantiated object.