Fixed: Ignored URL queries no longer removed from URL
The configurable ignored URL queries are intended to allow HashOver to treat two different URLs as being the same page. However, the URL queries were also being removed from the page URL HashOver uses to direct users back to the page they came from after logging in or deleting a comment, among other things.

This is obviously bad when the URL query being ignored is also important. For example, if a page uses a URL query to change the site language or theme, and that query is configured to be ignored so the that pages get the same comments, HashOver would redirect users back to a URL without those queries, thus removing the set language or theme.

This is now fixed, ignored URL queries are now only used to generate the same automatic thread name.