Added multisite support. (Using SQL? Read this commit)
HashOver can now be set to treat URLs from different websites as separate collections of threads, ensuring that two similar URLs from two different websites that get converted to the same thread name will not conflict with each other.

I am calling this "Multisite" support in the code, but it is not related to Wordpress. It might work on a Wordpress multisite setup, but it is not related to that functionality.

When HashOver is storing comments in flat file formats, like XML and JSON, the threads are now separated by being in different directories. One directory under `hashover/comments` for each unique domain name. If you enable multisite support you will therefore need to move the `hashover/comments/threads` and `hashover/comments/metadata` directories under a new directory named after your website domain name.

To support multiple websites in SQL databases, like SQLite and MySQL, would have required using more unconventional methods like creating multiple databases/sqlite files for each domain or further prefixing the table names with each domain. This was obviously not a good idea.

For this reason, the SQL database code has been completely rewritten. Comments are now stored as rows in a single `comments` table, page info metadata is stored as rows in a single `page-info` table, with each of these tables including columns for the website domain name and thread name which are used to query the right comments for the right website from the database.

This change will break existing HashOver SQL databases. SQL has always been a work in progress, so I hope not too many people were using SQL for an important website, as migrating those comments will now be necessary.

I have not yet figured out the easiest way to move comments from old SQL databases to the new format, but I hope it will be possible to do with a single SQL statement.

If you are using SQL, please open an issue saying so or contribute to an existing issue to help find a simple migration solution.

This change also happens to address issue #247.