Moved all files out of `hashover` directory
Having the files nested in a `hashover` directory was done to mirror the downloadable ZIP file to provide a standard install. However, in practice, having the files nested in a subdirectory has caused confusion. Cloning the files and then uploading the `hashover` directory to a web server is not how most people use Github, instead most opt to clone HashOver to their web server and use it in plase, as-is. This change makes sure HashOver is not nested after cloning in preparation for an upcoming commit.