Undo 4201fae, removes nested directory auto-detect support
After much testing, it appears there is no way to reliably detect the path of HashOver and create a relative path for HTTP use when HashOver is nested in one or more directories.

Generally, HashOver tries to "just work" and allow additional configuration where it doesn't. This is one of those cases. Many servers use symlinks in the document root path which cause conflicts with how HashOver resolves its path. To solve this problem I am opting to remove auto-detection of a nested relative path, but to support nested installations through manual configuration. Essentially the opposite of how it worked before. If you are using a standard install, this should not affect you.

Closes #211
Closes #251
Closes #256