Pass frontend settings by individual URL queries
Settings passed to HashOver during frontend instantiation are now passed via individual URL queries instead of one long settings query containing JSON data.

JSON data is quite small, that is why it was used, however, because of the necessity of encoding the JSON data it is actually quite inefficient. For example, the default settings passed to HashOver from the admin moderation page results in a URL that is 79 characters shorter than before, a 30% reduction.

With this new approach, we can pass more settings before hitting the URL length limit, view each query in the browser console, and avoid unnecessary JSON decoding on the server-side.