Added new session login method
You could already set the cookie expiration date to 'session' to have cookies expire at the end of the session, but this method uses an actual session to store the login information. This is slightly more secure than the cookies login, because the login information is kept on the server rather than on the user's computer.