Fixed bug where email field was automatically filled in edit forms, addresses #263
This was a long standing bug. Basically, when a user edits a comment the email address field was prefilled with the email of the current login, this is fine for normal users, but when admin edits a comment it was filled with the admin email. This being fixed also fixes a number of other issues, such as admin not being able to edit a user's email, and users being automatically re-logged-in when they edit a comment. This commit also removes a bit of code that automatically set the password field blank, as this too was added to address issues with the email field being prefilled.