Fixed new comments getting posted out of sort order
New comments are now added to the under-the-hood comment storage in canonical order, the comments are then sorted and (re)appended to the page. This ensures new comments appear in the correct place no matter what sorting is selected.

For example, if you were to select the "Newest First" sorting order, and then post a comment, that comment would be added at the bottom of the comments, which would make it appear to be the first comment ever posted according to the sorting order. This is no longer the case, the comment would now appear at the top.

This is also an example of a great commit. It expands functionality while reducing the codebase.