Removed user login and cookie restrictions from Admin
The `setsCookies` and `allowsLogin` settings are intended to only apply to normal users, not Admin, but before now they applied to both. This meant that with any configuration that caused either cookies or the user login to be disabled that Admin would not be able to login either. This was likely the cause of issue #273.