Improved how paths are concatenated
This helps prevent multiple forward slashes from appearing in paths, like "/hashover/themes//default/comments.css". This also fixes an issue most users are unlikely to encounter, where when the HashOver HTTP root is `/` (when HashOver is used as the root on a subdomain, for example) many paths would start with two forward slashes, like "//themes/default/comments.css" which is a valid path but points to the complete wrong place.