Ensure configured timezone is in ICU format, closes #278
For some reason the capitalization of a timezone ID does not matter for setting the default timezone with `date_default_timezone_set`, but it does matter for IntlDateFormatter::create. In fact, in the latter case the method simply fails without errors or exceptions. So the timezone now gets converted to ICU-style title case after validating it but before using it for Intl functions.