Added control over notification emails
You may now set notification emails to be sent to only Admin, only users when they reply to each other, to both Admin and users, or to nobody to disable notification emails. This new setting is available to the front-end, so if you want to disable email notifications for a specific page, you can. I now do this for the HashOver test pages.