Improved frontend settings and instantiation
Frontend settings are now passed to backend scripts where necessary. This fixes a bug where comments that were loaded after the initial interface would retain the default settings rather than the settings from the frontend, for example if you set the language to Japanese in the frontend and had comment collaping enabled, the comments that were collapsed would be displayed in the default language rather than Japanese, simply because the backend wasn't aware of the frontend setting changes.

Before now, the frontend would only include JavaScript code that was necessary for the current settings, this is no longer the case, all code is always included. This fixes a bug where if you disabled a feature from the frontend settings, and then enabled the same feature in a different instance from the frontend settings, the code would throw errors about undefined methods or simply retain the behavior expected of the default settings.

Before now, the theme HTML was loaded only once, with the first instance, this is no longer the case, the theme HTML is now loaded for each instance. This fixes a bug where if you changed the theme from the frontend settings in any instance but the first the comments would use the HTML of the first theme, rather than the one being set for the current instance.